requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyGithub>=2.1.1
pydantic>=2.0.0

# Web scraping and automation
selenium>=4.15.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and text processing
spacy>=3.7.0
nltk>=3.8.1

# Documentation and formatting
mkdocs>=1.5.0
mkdocs-material>=9.4.0
black>=23.10.0
flake8>=6.1.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Type checking
mypy>=1.6.0
types-requests>=2.31.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Testing dependencies
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
coverage>=7.3.0
aioresponses>=0.7.4
async-timeout>=4.0.0

# Development dependencies
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# Monitoring UI dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
psutil>=5.8.0

# Performance optimization
cachetools>=5.3.0
aiofiles>=23.1.0 